Refactor namespace structure: relocate `MessageHandlers` into `Interop` for improved organization
