+++ examples/InfiniFrameExample.WebApp.Blazor/Program.cs
@@ -3,11 +3,11 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using InfiniFrame;
 using InfiniFrame.Js;
-using InfiniFrame.Js.MessageHandlers;
 using InfiniFrame.WebServer;
 using InfiniFrameExample.WebApp.Blazor.Components;
 using Serilog;
 using System.Drawing;
+using InfiniFrame.Js.Interop.MessageHandlers;
 
 namespace InfiniFrameExample.WebApp.Blazor;
 // ---------------------------------------------------------------------------------------------------------------------
