+++ examples/InfiniFrameExample.WebApp.Vue/Program.cs
@@ -2,9 +2,9 @@
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
 using InfiniFrame;
-using InfiniFrame.Js.MessageHandlers;
 using InfiniFrame.WebServer;
 using System.Drawing;
+using InfiniFrame.Js.Interop.MessageHandlers;
 
 namespace InfiniFrameExample.WebApp.Vue;
 // ---------------------------------------------------------------------------------------------------------------------
