+++ src/InfiniFrame.Js/Interop/MessageHandlers/TitleChangedWebMessageHandler.cs
@@ -3,7 +3,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using Microsoft.Extensions.Logging;
 
-namespace InfiniFrame.Js.MessageHandlers;
+namespace InfiniFrame.Js.Interop.MessageHandlers;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
