+++ src/InfiniFrame.Js/Interop/RegisterWindowCreatedUtility.cs
@@ -2,10 +2,9 @@
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
 using System.Runtime.CompilerServices;
-using InfiniFrame.Js.Interop;
 using Microsoft.Extensions.Logging;
 
-namespace InfiniFrame.Js.MessageHandlers;
+namespace InfiniFrame.Js.Interop;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
