+++ src/InfiniFrame.Js/Interop/WindowReadyRegistrationState.cs
@@ -3,7 +3,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using System.Runtime.CompilerServices;
 
-namespace InfiniFrame.Js.MessageHandlers;
+namespace InfiniFrame.Js.Interop;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
